Workaround to ensure GitHub actions will work
Currently importing datetime2 fails, because western time doesn't have
the from_time_pair method which is required for registration. I cannot
implement it on the fly, because it looks also that documentation doesn't
cover it. I will have to recover the normal development loop (doc, tests,
code) immediately after the new build tools are in place.